refactor: Splits out TestTracer
